Moved shared cypress stub http calls code to support/helpers. Added timeout when shutting down terminal workers for model registration.
